# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Employee Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leave-types:
    get:
      operationId: listLeaveTypes
      tags: [leaves]
      summary: List leave types
      responses:
        "200":
          description: List of leave types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveType"
    post:
      operationId: createLeaveType
      tags: [leaves]
      summary: Create leave type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveType"
  /leave-types/{id}:
    patch:
      operationId: updateLeaveType
      tags: [leaves]
      summary: Update leave type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveTypeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveType"
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: leaveTypeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leave"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Submit leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/balance:
    get:
      operationId: getLeaveBalance
      tags: [leaves]
      summary: Get leave balance for an employee
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveBalance"
  /leaves/{id}:
    get:
      operationId: getLeave
      tags: [leaves]
      summary: Get leave request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
    patch:
      operationId: updateLeave
      tags: [leaves]
      summary: Update leave request (approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
    delete:
      operationId: deleteLeave
      tags: [leaves]
      summary: Delete leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /payroll:
    get:
      operationId: listPayroll
      tags: [payroll]
      summary: List payroll records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payroll records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
    post:
      operationId: createPayroll
      tags: [payroll]
      summary: Generate payroll for period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/summary:
    get:
      operationId: getPayrollSummary
      tags: [payroll]
      summary: Payroll summary
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payroll summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSummary"
  /payroll/{id}:
    get:
      operationId: getPayroll
      tags: [payroll]
      summary: Get payroll by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
    patch:
      operationId: updatePayroll
      tags: [payroll]
      summary: Update payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
    delete:
      operationId: deletePayroll
      tags: [payroll]
      summary: Delete payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LeaveType:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        maxDaysPerYear:
          type: integer
        isPaid:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveTypeInput:
      type: object
      required: [name, maxDaysPerYear]
      properties:
        name:
          type: string
        maxDaysPerYear:
          type: integer
        isPaid:
          type: boolean
        description:
          type: string
    LeaveTypeUpdate:
      type: object
      properties:
        name:
          type: string
        maxDaysPerYear:
          type: integer
        isPaid:
          type: boolean
        description:
          type: string
    Leave:
      type: object
      required: [id, employeeId, leaveTypeId, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        totalDays:
          type: integer
        reason:
          type: ["string", "null"]
        status:
          type: string
        managerApprovalStatus:
          type: ["string", "null"]
        hrApprovalStatus:
          type: ["string", "null"]
        managerNote:
          type: ["string", "null"]
        hrNote:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveInput:
      type: object
      required: [leaveTypeId, startDate, endDate]
      properties:
        employeeId:
          type: ["integer", "null"]
        leaveTypeId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    LeaveUpdate:
      type: object
      properties:
        status:
          type: string
        managerApprovalStatus:
          type: string
        hrApprovalStatus:
          type: string
        managerNote:
          type: string
        hrNote:
          type: string
        reason:
          type: string
    LeaveBalance:
      type: object
      required: [leaveTypeId, leaveTypeName, totalAllowed, used, remaining]
      properties:
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: string
        totalAllowed:
          type: integer
        used:
          type: integer
        remaining:
          type: integer
...
    Payroll:
      type: object
      required: [id, employeeId, month, year, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        employeeCode:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        allowances:
          type: number
        lateDeductions:
          type: number
        leaveDeductions:
          type: number
        otherDeductions:
          type: number
        netSalary:
          type: number
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        status:
          type: string
        createdAt:
          type: string
    PayrollInput:
      type: object
      required: [employeeId, month, year]
      properties:
        employeeId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        allowances:
          type: number
        otherDeductions:
          type: number
    PayrollUpdate:
      type: object
      properties:
        allowances:
          type: number
        otherDeductions:
          type: number
        status:
          type: string
    PayrollSummary:
      type: object
      required: [totalEmployees, totalGross, totalDeductions, totalNet, month, year]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalEmployees:
          type: integer
        totalGross:
          type: number
        totalDeductions:
          type: number
        totalNet:
          type: number
        byDepartment:
          type: array
          items:
            type: object
            required: [department, totalNet]
            properties:
              department:
                type: string
              totalNet:
                type: number
...