# lib/api-zod/src/generated/types/leaveType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveType {
  id: number;
  name: string;
  maxDaysPerYear?: number;
  isPaid?: boolean;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}