# lib/api-zod/src/generated/types/payrollInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PayrollInput {
  employeeId: number;
  month: number;
  year: number;
  basicSalary?: number;
  allowances?: number;
  otherDeductions?: number;
}